Enhancements to provide more legancy on solving issue #7
Now a type can be configured recursively. That is, you provide an
aggregate root and all associations, associations of associations and so
on get configured recursively. A filter can be provided to avoid
tracking all properties. Since this is a work in progress feature, a
future commit will provide documentation on this...